+++ packages/liveblocks-react/src/types.ts
@@ -27,7 +27,8 @@ import type {
   DeleteCommentOptions,
   EditCommentOptions,
   EditThreadMetadataOptions,
-  RoomThreads,
+  ThreadsState,
+  ThreadsStateSuccess,
 } from "./comments/CommentsRoom";
 
 export type ResolveUserOptions = {
@@ -39,26 +40,28 @@ export type ResolveMentionSuggestionsOptions = {
   text: string;
 };
 
-export type UserState<T> =
-  | {
-      user?: never;
-      isLoading: true;
-      error?: never;
-    }
-  | {
-      user?: T;
-      isLoading: false;
-      error?: never;
-    }
-  | {
-      user?: never;
-      isLoading: false;
-      error: Error;
-    };
+export type UserStateLoading = {
+  isLoading: true;
+  user?: never;
+  error?: never;
+};
 
-export type UserStateSuspense<T> = Resolve<
-  Extract<UserState<T>, { isLoading: false; error?: never }>
->;
+export type UserStateError = {
+  isLoading: false;
+  user?: never;
+  error: Error;
+};
+
+export type UserStateSuccess<T> = {
+  isLoading: false;
+  user: T;
+  error?: never;
+};
+
+export type UserState<T> =
+  | UserStateLoading
+  | UserStateError
+  | UserStateSuccess<T>;
 
 export type RoomProviderProps<
   TPresence extends JsonObject,
@@ -602,7 +605,7 @@ export type RoomContextBundle<
      * @example
      * const { threads, error, isLoading } = useThreads();
      */
-    useThreads(): RoomThreads<TThreadMetadata>;
+    useThreads(): ThreadsState<TThreadMetadata>;
 
     /**
      * @beta
@@ -750,9 +753,9 @@ export type RoomContextBundle<
          * Returns the threads within the current room.
          *
          * @example
-         * const threads = useThreads();
+         * const { threads } = useThreads();
          */
-        useThreads(): ThreadData<TThreadMetadata>[];
+        useThreads(): ThreadsStateSuccess<TThreadMetadata>;
 
         /**
          * @beta
@@ -760,9 +763,9 @@ export type RoomContextBundle<
          * Returns user info from a given user ID.
          *
          * @example
-         * const { user, error, isLoading } = useUser("user-id");
+         * const { user } = useUser("user-id");
          */
-        useUser(userId: string): UserStateSuspense<TUserMeta["info"]>;
+        useUser(userId: string): UserStateSuccess<TUserMeta["info"]>;
 
         //
         // Legacy hooks